add new syscalls and refactor reading state
