+++ api/src/index.ts
@@ -6,7 +6,6 @@ export { GearProgram } from './Program';
 export { GearMessage } from './Message';
 export { GearMessageReply } from './MessageReply';
 export { GearBalance } from './Balance';
-export { getWasmMetadata } from './WasmMeta';
 export { DebugMode } from './DebugMode';
 export { GearMailbox } from './Mailbox';
 export { GearProgramState } from './State';
@@ -14,6 +13,7 @@ export { GearBlock } from './Blocks';
 export { GearGasSpent } from './GasSpent';
 export { GearStorage } from './Storage';
 export { GearClaimValue } from './Claim';
+export * from './wasm';
 export * from './CreateType';
 export * from './create-type';
 export * from './interfaces';
