+++ api/src/wasm/WasmMeta.ts
@@ -1,0 +1,27 @@
+import { Metadata } from '../interfaces';
+import importObj from './importObj';
+import { readMetaValue } from './utils';
+
+export async function getWasmMetadata(wasmBytes: Buffer, showDebug = false): Promise<Metadata> {
+  const memory = new WebAssembly.Memory({ initial: 256 });
+  let module = await WebAssembly.instantiate(wasmBytes, importObj(memory, showDebug));
+  const exports = module.instance.exports;
+  if (!exports) {
+    return {};
+  }
+  let metadata: Metadata = {
+    types: `0x${readMetaValue(memory, exports.meta_registry)}`,
+    init_input: readMetaValue(memory, exports.meta_init_input),
+    init_output: readMetaValue(memory, exports.meta_init_output),
+    async_init_input: readMetaValue(memory, exports.meta_async_init_input),
+    async_init_output: readMetaValue(memory, exports.meta_async_init_output),
+    handle_input: readMetaValue(memory, exports.meta_handle_input),
+    handle_output: readMetaValue(memory, exports.meta_handle_output),
+    async_handle_input: readMetaValue(memory, exports.meta_async_handle_input),
+    async_handle_output: readMetaValue(memory, exports.meta_async_handle_output),
+    title: readMetaValue(memory, exports.meta_title),
+    meta_state_input: readMetaValue(memory, exports.meta_state_input),
+    meta_state_output: readMetaValue(memory, exports.meta_state_output),
+  };
+  return metadata;
+}
